solved ticket https://issues.jboss.org/browse/RAILO-2675

+++ railo-java/railo-core/src/railo/runtime/converter/JSConverter.java
@@ -88,7 +88,7 @@ public final class JSConverter extends ConverterSupport {
 		if(object instanceof Number) {
 			sb.append(goIn());
 			sb.append("\"");
-			sb.append(Caster.toString(((Number)object).doubleValue()));
+			sb.append(Caster.toString(((Number)object)));
 			sb.append("\";");
 			return;
 		}
